+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * When saving yaml and markdown, create also a cached version of the file and recompile it in opcache
 2. [](#bugfix)
     * Fixed missing changes in yaml & markdown files if saved multiple times during the same second because of a caching issue
+    * Fixed XSS check not detecting onX events without quotes
 
 # v1.7.32
 ## 03/28/2022